[MER-5677] Isolate Phases of the Moon logic from the shared adaptive driver
The Moon-specific extensions to AdaptiveDeckPO.ts/AdaptiveHappyPathTask.ts
(modal close, Moon simulator, phase sorting, image-clarity MCQ, reflector
FITB) previously ran unconditionally for every adaptive lesson test,
including real-chem-dazzling-d-orbitals and real-chem-greenhouse-molecules,
which also share these files. Move all lesson-specific logic into a new
tasks/lessons/PhasesOfTheMoonTask.ts extension, invoked from a single call
site only when the lesson's answers declare moon/reflector_fitb/clarity_mcq.

Also genericize scanScreen/fillTextInputs with an opt-in
extra_text_input_selectors field instead of unconditionally widening the
text-input selector to include textarea for every lesson. Document both
shared files' extension conventions for future lessons.
